+++ examples/nextjs/__tests__/locations/EntryEditor.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { mockCma, mockSdk } from '../../__tests__/mocks';
-import EntryEditor from '@/pages/locations/EntryEditor';
+import EntryEditor from '@/components/locations/EntryEditor';
 
 jest.mock('@contentful/react-apps-toolkit', () => ({
   useSDK: () => mockSdk,
