+++ examples/nextjs/components/locations/ConfigScreen.tsx
