+++ examples/nextjs/components/locations/EntryEditor.tsx
