+++ examples/nextjs/components/locations/Sidebar.tsx
